댓글 작성 조희 삭제 기능

+++ post_CRUD/src/main/java/com/sparta/post_crud/service/ReviewService.java
@@ -17,24 +17,13 @@ public class ReviewService {
     private final ReviewRepository reviewRepository;
 
     @Transactional
-    public ResponseDto<?> createReview(ReviewRequestDto requestDto) {
+    public ResponseDto<?> createReview(ReviewRequestDto requestDto,String username) {
 
-        Review Review = new Review(requestDto);
+        Review Review = new Review(requestDto, username);
 
         reviewRepository.save(Review);
 
         return ResponseDto.success(Review);
-    }
-
-    @Transactional(readOnly = true)
-    public ResponseDto<?> getReview(Long id) {
-        Optional<Review> optionalReview = reviewRepository.findById(id);
-
-        if (optionalReview.isEmpty()) {
-            return ResponseDto.fail("NULL_Review_ID", STR_NULL);
-        }
-
-        return ResponseDto.success(optionalReview.get());
     }
 
     @Transactional(readOnly = true)
@@ -43,7 +32,7 @@ public class ReviewService {
     }
 
     @Transactional
-    public ResponseDto<Review> updateReview(Long id, ReviewRequestDto requestDto) {
+    public ResponseDto<Review> updateReview(Long id, ReviewRequestDto requestDto , String username) {
         Optional<Review> optionalReview = reviewRepository.findById(id);
 
         if (optionalReview.isEmpty()) {
@@ -51,13 +40,17 @@ public class ReviewService {
         }
 
         Review review = optionalReview.get();
-        review.update(requestDto);
+        if(!review.getUsername().equals(username)){
+            return ResponseDto.fail("NOT_REVIEWING_USER","댓글 작성자가 아닙니다.");
+        }
+
+        review.update(requestDto,username);
 
         return ResponseDto.success(review);
     }
 
     @Transactional
-    public ResponseDto<?> deleteReview(Long id) {
+    public ResponseDto<?> deleteReview(Long id,String username) {
         Optional<Review> optionalReview = reviewRepository.findById(id);
 
         if (optionalReview.isEmpty()) {
@@ -65,9 +58,10 @@ public class ReviewService {
         }
 
         Review review = optionalReview.get();
-
+        if(!review.getUsername().equals(username)){
+            return ResponseDto.fail("NOT_REVIEWING_USER","댓글 작성자가 아닙니다.");
+        }
         reviewRepository.delete(review);
-
         return ResponseDto.success(true);
     }
